net: ks8851: Fixed ks8851_rdreg issue on MXS arch
i.MX23 and i.MX28 has a half-duplex chipset (Sigmatel SPI block),
which means ks8851_rdreg doesn't work out of the box. At the same
time ks8851 driver has a selector ks8851_rx_1msg that can be used
as a w/a for chipsets that cannot be used in a full-duplex mode.
In original implementation this method return always true, but
this patch introduce more arch test and return false for MXS arch.

Signed-off-by: Sergey Shcherbakov <shchers@gmail.com>